piperci-flake8-gateway/requirements.txt
requests
PyYAML
pipercilib@git+https://github.com/afcyber-dream/python-piperci.git
